moved dispatches to App.js
